refac: news cards
